+++ lib/Completion/Bridge/TolerantParser/ReferenceFinder/AttributeCompletor.php
@@ -4,41 +4,28 @@ namespace Phpactor\Completion\Bridge\TolerantParser\ReferenceFinder;
 
 use Generator;
 use Microsoft\PhpParser\Node;
+use Microsoft\PhpParser\Node\QualifiedName;
 use Phpactor\Completion\Bridge\TolerantParser\CompletionContext;
 use Phpactor\Completion\Bridge\TolerantParser\TolerantCompletor;
-use Phpactor\Completion\Core\DocumentPrioritizer\DocumentPrioritizer;
-use Phpactor\Completion\Core\Suggestion;
-use Phpactor\ReferenceFinder\NameSearcher;
-use Phpactor\ReferenceFinder\NameSearcherType;
+use Phpactor\Completion\Core\Completor\NameSearcherCompletor;
+use Phpactor\Name\NameUtil;
 use Phpactor\TextDocument\ByteOffset;
 use Phpactor\TextDocument\TextDocument;
 
-class AttributeCompletor implements TolerantCompletor
+class AttributeCompletor extends NameSearcherCompletor implements TolerantCompletor
 {
-    public function __construct(
-        private NameSearcher $nameSearcher,
-        private DocumentPrioritizer $prioritizer
-    ) {
-    }
-
     public function complete(Node $node, TextDocument $source, ByteOffset $offset): Generator
     {
         if (!CompletionContext::attribute($node)) {
             return true;
         }
 
-        $search = $node->getText();
-        $type = NameSearcherType::CLASS_;
-
-        foreach ($this->nameSearcher->search($search, $type) as $result) {
-            yield Suggestion::createWithOptions($result->name()->head(), [
-                'type' => Suggestion::TYPE_CLASS,
-                'priority' => $this->prioritizer->priority($result->uri(), $source->uri()),
-                'short_description' => sprintf('%s %s', $type, $result->name()->__toString()),
-                'class_import' => $result->name()->__toString(),
-                'name_import' => $result->name()->__toString(),
-            ]);
+        $name = $node->__toString();
+        if ($node instanceof QualifiedName && NameUtil::isQualified($name)) {
+            $name = NameUtil::toFullyQualfiied((string)$node->getResolvedName());
         }
+
+        yield from $this->completeName($name, $source->uri(), $node);
 
         return true;
     }