+++ lib/Completion/Bridge/TolerantParser/WorseReflection/DoctrineAnnotationCompletor.php
@@ -10,12 +10,14 @@ use Phpactor\Completion\Core\Completor;
 use Phpactor\Completion\Core\Completor\NameSearcherCompletor;
 use Phpactor\Completion\Core\Suggestion;
 use Phpactor\Completion\Core\Util\OffsetHelper;
+use Phpactor\Name\NameUtil;
 use Phpactor\ReferenceFinder\NameSearcher;
 use Phpactor\ReferenceFinder\Search\NameSearchResult;
 use Phpactor\TextDocument\ByteOffset;
 use Phpactor\TextDocument\TextDocument;
 use Phpactor\TextDocument\TextDocumentUri;
 use Phpactor\WorseReflection\Core\Exception\NotFound;
+use Phpactor\WorseReflection\Core\Util\NodeUtil;
 use Phpactor\WorseReflection\Reflector;
 
 class DoctrineAnnotationCompletor extends NameSearcherCompletor implements Completor
@@ -50,6 +52,11 @@ class DoctrineAnnotationCompletor extends NameSearcherCompletor implements Compl
         if (!$annotation = $this->extractAnnotation($truncatedSource)) {
             // Ignore if not an annotation
             return true;
+        }
+
+        $namespace = NodeUtil::namespace($node);
+        if (NameUtil::isQualified($annotation) && $namespace) {
+            $annotation = '\\' . NameUtil::join($namespace, $annotation);
         }
 
         $suggestions = $this->completeName($annotation, $source->uri());